sistema-pishing/requirements.txt
streamlit
pandas
plotly
requests
fastapi
uvicorn
supabase
pandas
requests
python-dotenv
asyncpg
fpdf2
